update
Adding change log entries.
Found the culprit with the transform ownership instability. Moving an object can give it velocity which if there is enough time that passes between updates then fixed update will apply enough velocity motion to move the object and it will continue to move.
Increased mass and linearDampening and setting the linear velocity to zero in the appropriate places.