chore: merge release 2.2.0 back into develop-2.0.0 (#3179)
* update

updating package

* update

updating changelog

* update

moving minimum editor back down to 6000.0

* fix

Fixing the changelog version goof of mine.
Making the license copyright just a year and not year range.

* style

removing whitespace from copyright notice.
adding dash to changelog header

* style

Removing whitespaces for PVP-124-2 issues.

* style

missed one whitespace...

* chore: merge develop 2 0 0 updates with SessionOwner permissions (#3176)

* fix: clamp spawntimeout to recommended range (#3174)

* update

Clamping spawntimeout

* update

improving parenting failed message when either the child or parent NetworkObject is not spawnd.

* update

Update the local SceneEventData.SceneEventType on the authority side for SceneLoadComplete.

* style

removing whitespaces

* feat: Add a SessionOwner ObjectStatus and allow InScenePlaced network objects to be distributed (#3175)

* Initial pass on SessionOwner ownership flag

* feat: Add SessionOwner OwnershipStatus flag

* update

removing additional session owner accessor.

* Add OwnershipPermissions tests

* fix client connect

* Revert "fix client connect"

This reverts commit 3c3b35444200da308dd07e1dd2e7b591fa0842d5.

* update

object distribution for in-scene placed NetworkObjects needs to use the InScenePlacedSourceGlobalObjectIdHash when building an object distribution list.

* Add changelog

* Remove unnecessary change

* Remove Settings.json

* Reword CHANGELOG

* fix

DAHost should not distribute session owner permission NetworkObjects.
When client is promoted to session owner, for now newly promoted client takes ownership of NetworkObjects that have the SessionOwner permission set.
Only prevent non-session owners from taking ownership of a NetworkObject with the SessionOwner permission set.

* test fix

Avoid the RemoveOwnership client-server only method.

* style

Visual studio code cleanup likes to sort by alpha... fixing for our standards.

* update

Adding check for session owner trying to change ownership to a non-session owner client.

* test

Adding an additional validation that a non-session owner cannot change ownership and that a session owner cannot change ownership to a non-session owner when the NetworkObject in question has the SessionOwner permissions set.

---------

Co-authored-by: NoelStephensUnity <noel.stephens@unity3d.com>

---------

Co-authored-by: Emma <emma.mcmillan@unity3d.com>

* fix

NetworkAnimator started throwing this exception:
"Objects are trying to be loaded during a domain backup. This is not allowed as it will lead to undefined behaviour!"
When MPPM was being used. This has to do with the OnBeforeSerialize method and potentially an order of operations issue.
We can accomplish the same serialization within OnValidate, but went ahead and made it virtual in order to provide users with the ability to override and handle their own validation.

* fix: server, host, or session owner not populating in-scene placed object table when started (#3177)

* fix

This fixes the issue with the server, host, or session owner not adding in-scene placed NetworkObjects to its internal table if the scene was loaded prior to starting the NetworkManager.

* update

Adding change log entry

* update

Adding PR# to this entry.

* style

remove whitespace

* update

adding unreleased change log

---------

Co-authored-by: Emma <emma.mcmillan@unity3d.com>